feat(datasource): allow the module to test the Prometheus datasource after creation/update
The Grafana WUI as a feature on datasource that allows to 'Save & test' the datasource
when creating or updating it.

This patch allows to run the same query after Prometheus datasource
creation and update.

A GrafanaHttpError exception was added to improve error management and
returned message in case of failure.