Replace GHA boilerplate with the use of an Action
This patch replaces manually maintained copy-paste of `ansible-test`
invocations with the use of a reusable GitHub Action[1] that wraps
it allowing to only keep the matrix on the project side allowing the
reusable bits to be maintained separately and be shared across the
community.

This has been earlier implemented for the `community.digitalocean`
collection[2] among others.

[1]: https://github.com/ansible-community/ansible-test-gh-action
[2]: https://github.com/ansible-collections/community.digitalocean/pull/144