Use ansible-test-gh-action GHA @ integration tests
Supersedes
https://github.com/ansible-collections/community.grafana/pull/266.

This patch replaces manually maintained copy-paste of `ansible-test`
invocations with the use of a reusable GitHub Action[[1]] that wraps
it allowing to only keep the matrix on the project side allowing the
reusable bits to be maintained separately and be shared across the
community.

This has been earlier implemented for the `community.digitalocean`
collection[[2]] and some tests in this collection[[3]] among others.

Resolves #266

[1]: https://github.com/ansible-community/ansible-test-gh-action
[2]: https://github.com/ansible-collections/community.digitalocean/pull/144
[3]: https://github.com/ansible-collections/community.grafana/pull/270