fix(datasource-elastic): allow support of elastic version > 7
In latest versions of Grafana, the elastic version is no more provided
as an integer even if it is still supported and won't return any error.
To ease configuration, the module will accept the version format as
displayed by the WUI (ex: "7.10+") and will map the value to what the API
expects ("7.10.0").
The added values are the one supported by Grafana 8 and 9 (greater than
elastic 7.0 which was already supported).

The module will keep on supporting the old format to avoid a breaking
change with significant impact.

Fixes: #263